requirements.txt
argon2-cffi==20.1.0
watchgod==0.7
uvicorn[standard]

django==3.2.7
django-environ
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2
pytest
pytest-django
pytest-cov
factory-boy